requirements.txt
# Core dependencies
fastapi>=0.109.0,<0.110.0
uvicorn[standard]>=0.27.0,<0.28.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-multipart>=0.0.6

# NLP & Text Processing
spacy>=3.7.0,<4.0.0
langdetect>=1.0.9
nltk>=3.8.1

# Retrieval & Embeddings
sentence-transformers>=2.3.0,<3.0.0
faiss-cpu>=1.7.0,<2.0.0
rank-bm25>=0.2.2

# LLM - Install PyTorch first separately if needed
transformers>=4.36.0,<5.0.0
torch>=2.1.0,<3.0.0
accelerate>=0.25.0,<1.0.0
# bitsandbytes - REMOVED: Not compatible with Windows

# Anomaly Detection & ML - Using flexible versions for better Windows compatibility
scikit-learn>=1.3.0,<2.0.0
numpy>=1.24.0,<2.0.0
pandas>=2.0.0,<3.0.0
scipy>=1.11.0,<2.0.0

# Data Processing
python-dotenv>=1.0.0
pyyaml>=6.0

# Database & Storage
elasticsearch>=8.11.0,<9.0.0

# Logging & Monitoring
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
httpx>=0.26.0
requests>=2.31.0

# Terminal Output (for scenario tests)
colorama>=0.4.6

# Development
black>=23.12.0
flake8>=7.0.0
mypy>=1.8.0


